Refactor: Banner 컴포넌트 분리 PostList 컴포넌트 분리 :
