+++ src/App.jsx
@@ -3,7 +3,7 @@ import GlobalStyle from './GlobalStyle';
 import Navigation from './components/Navigation';
 import Home from './pages/Home';
 import Login from './pages/Login';
-import Mypage from './pages/Mypage/Mypage';
+import Mypage from './pages/Mypage';
 import { Route, Routes } from 'react-router-dom';
 import { useEffect } from 'react';
 import { onAuthStateChanged } from 'firebase/auth';
